svd df/f change
changed the dffFromSVD function:

1) changed the df/f to soft norm because zeros in the denominator caused all NaNs and low values blew up

2) more importantly: added a ChangeU call at the end of the function to but the df/f V's into the original U-space

I did this because I was attempting to use a common U-space across experiments, but I was getting weird changes in resulting signal amplitude. The reason was that the new U's after df/f-ing are not orthonormal, which is a requirement for a lot of operations that happen in V-space.